Origin, Destination complete. Missing POST

+++ app/src/main/java/cit/jauc/lib/CoordsConverter.java
@@ -2,7 +2,7 @@ package cit.jauc.lib;
 
 public class CoordsConverter {
 
-    public static String getLocationfromCoords(long longitude, long latitude) {
+    public static String getLocationfromCoords(double longitude, double latitude) {
         // TODO API Call to get location from coords
         return "Location " + longitude + "," + latitude;
     }
