+++ app/src/main/java/cit/jauc/model/Location.java
@@ -3,31 +3,31 @@ package cit.jauc.model;
 import java.io.Serializable;
 
 public class Location implements Serializable {
-    private long lon;
-    private long lat;
+    private double lon;
+    private double lat;
 
     public Location() {
 
     }
 
-    public Location(long lon, long lat) {
+    public Location(double lon, double lat) {
         this.lon = lon;
         this.lat = lat;
     }
 
-    public long getLon() {
+    public double getLon() {
         return lon;
     }
 
-    public void setLon(long lon) {
+    public void setLon(double lon) {
         this.lon = lon;
     }
 
-    public long getLat() {
+    public double getLat() {
         return lat;
     }
 
-    public void setLat(long lat) {
+    public void setLat(double lat) {
         this.lat = lat;
     }
